refactor: member controller에 logout endpoint 구현
