+++ src/main/java/hyundai/movie_review/member/controller/MemberController.java
@@ -15,6 +15,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -119,5 +122,24 @@ public class MemberController {
         return ResponseEntity.ok(response);
     }
 
+    @Operation(description = "사용자 로그아웃 엔드포인트")
+    @PostMapping("/member/logout")
+    public void logout(HttpServletRequest request, HttpServletResponse response) {
+        // Access Token 삭제
+        invalidateCookie(response, "accessToken", request.getServerName());
+
+        // Refresh Token 삭제
+        invalidateCookie(response, "refreshToken", request.getServerName());
+    }
+
+    private void invalidateCookie(HttpServletResponse response, String name, String domain) {
+        Cookie cookie = new Cookie(name, null); // 쿠키 값을 null로 설정
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+        cookie.setDomain(domain.contains("localhost") ? "localhost" : "theaterup.site");
+        cookie.setSecure(!domain.contains("localhost")); // Secure 속성은 배포 환경에서만 활성화
+        cookie.setMaxAge(0); // 즉시 만료
+        response.addCookie(cookie);
+    }
 
 }
